server/requirements.txt
flask
flask-cors
yt-dlp
psycopg2-binary
gunicorn